% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.soilMoisture.R
\name{init.soilMoisture}
\alias{init.soilMoisture}
\title{Initialize information related to Soil Moisture}
\usage{
init.soilMoisture(method = NULL, path = NULL, waterSoil = NULL,
  waterGlaciatedSoil = NULL, waterGlaciers = NULL, Z = NULL,
  isoil = NULL, gisoil = NULL, bisoil = NULL, swgt = NULL,
  gwgt = NULL, snowfree = NULL, glacfrac = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual", "procecessed"}

\item{path}{directory where to get the files}

\item{waterSoil}{amount of water in soil (in mm)}

\item{waterGlaciatedSoil}{amount of water in glaciated soil (in mm)}

\item{waterGlaciers}{amount of water in glaciers (in mm)}

\item{Z}{waterVolume, input from rain, snow and glaciers (in mm)}

\item{isoil}{precipitation and/or snowmelt from the elevation zones, vector(level zone)}

\item{gisoil}{glaciermelt from the elevation zones, vector(level zone)}

\item{bisoil}{bog melt from the elevation zones, vector(level zone)}

\item{swgt}{weights for input for each elevation zone}

\item{gwgt}{weights for input for each elevation zone}

\item{snowfree}{weights for input for each elevation zone}

\item{glacfrac}{fraction of glaciers area}
}
\description{
Initialize the information related to ssoil Moisture
}
\examples{
\dontrun{
init.soilMoisture()
}
}
\keyword{soilMoisture}

