% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.soilMoisture.R
\name{stateX.soilMoisture}
\alias{stateX.soilMoisture}
\title{stateX.soilMoisture}
\usage{
stateX.soilMoisture(isoil, gisoil, bisoil = NULL, swgt, gwgt, snowfree,
  glacfrac)
}
\arguments{
\item{isoil}{precipitation and/or snowmelt from the elevation zones, vector(level zone)}

\item{gisoil}{glaciermelt from the elevation zones, vector(level zone)}

\item{bisoil}{bog melt from the elevation zones, vector(level zone)}

\item{swgt}{weights for input for each elevation zone}

\item{gwgt}{weights for input for each elevation zone}

\item{snowfree}{weights for input for each elevation zone}

\item{glacfrac}{fraction of glaciers area}
}
\description{
The function \code{stateX.soilMoisture()} provides the key features of the soil Moisture.
1- Compute the water amount (in mm) of precipitation and snowmelt, for each level zone:
i-on non-glaciated soils
ii- on glaciated soils
iii- on glaciers
iv- NOT YET on bogs
2- compute the Z(t): Actual water volume present in the soil moisture zone (Skaugen et Onof, 2013)
}
\examples{
\dontrun{
stateX()
}
}
\keyword{soilMoisture}

