% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amount.R
\name{waterAmount}
\alias{waterAmount}
\title{Amount of water}
\usage{
waterAmount(isoil, wgt)
}
\arguments{
\item{isoil}{precipitation and/or snowmelt for each of the elevation zones}

\item{wgt}{weights of the input for each of the elevation zones}
}
\description{
The function \code{waterAmount(} computes the amount of water
Unit is in mm
}
\examples{
\dontrun{
waterAmount()
}
}
\keyword{soilMoisture}

