% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterVolume.R
\name{waterVolume}
\alias{waterVolume}
\title{Volume of water}
\usage{
waterVolume(waterSoil, waterGlaciatedSoil, waterGlaciers, waterBogs = NULL,
  glacfrac)
}
\arguments{
\item{waterSoil}{amount of water on non glaciated soil}

\item{waterGlaciatedSoil}{amount of melting water on glaciated soil}

\item{waterGlaciers}{amount of melting water on glaciers}

\item{waterBogs}{amount of water in bogs. Not taken into account so far}

\item{glacfrac}{fraction of glaciers area}
}
\description{
The function \code{waterAmount(} computes the amount of water
Unit is in mm
}
\examples{
\dontrun{
waterVolume()
}
}
\keyword{soilMoisture}

